# [TTree] Avoid getting the number of branches twice