The operation is relatively costly as TObjArray::GetEntries also
creates a TReadLockGuard.