@bluehood 2019-03-26
Measurements for the benchmarks in the `philippe`  branch of the [df_bench](https://gitlab.com/bluehood/df_bench) repository.

Before patch:
```
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
TreeGetEntryOneBranch                        247377 ns       247373 ns         2856
TreeGetEntryTwoBranchesAndAllStatusFalse     310731 ns       310725 ns         2258
TreeGetEntryTwoBranches                      381067 ns       381058 ns         1828
TreeGetEntryTwoBranchesWithoutSetStatus      985630 ns       985614 ns          701
BranchGetEntryOneBranch                      103621 ns       103620 ns         6753
BranchGetEntryTwoBranches                    103768 ns       103764 ns         6749
```

After patch:
```
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
TreeGetEntryOneBranch                        180852 ns       180846 ns         3872
TreeGetEntryTwoBranchesAndAllStatusFalse     220974 ns       220957 ns         3180
TreeGetEntryTwoBranches                      292639 ns       292633 ns         2396
TreeGetEntryTwoBranchesWithoutSetStatus      900375 ns       900325 ns          773
BranchGetEntryOneBranch                      104265 ns       104262 ns         6592
BranchGetEntryTwoBranches                    105086 ns       105083 ns         6686
```